ParsaChangeEntitiesDirectoriesInOnion
