+++ App.Domain.AppServices/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace App.Domain.AppServices
+{
+    public class Class1
+    {
+
+    }
+}
