+++ App.Domain.Core/BaseData-Aggregate/Entities/Color.cs
@@ -1,9 +1,13 @@
-﻿using System;
+﻿using App.Domain.Core.Product_Aggregate.Entities;
+using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.BaseData_Aggregate.Entities
 {
-    public partial class Color
+    public class Color
     {
         public Color()
         {
