+++ App.Domain.Core/BaseData-Aggregate/Entities/FileType.cs
@@ -1,9 +1,10 @@
-﻿using System;
+﻿using App.Domain.Core.Product_Aggregate.Entities;
+using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.BaseData_Aggregate.Entities
 {
-    public partial class FileType
+    public class FileType
     {
         public FileType()
         {
