+++ App.Domain.Core/BaseData-Aggregate/Entities/FileTypeExtention.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.BaseData_Aggregate.Entities
 {
-    public partial class User
+    public class FileTypeExtention
     {
-        public User()
+        public FileTypeExtention()
         {
-            Comments = new HashSet<Comment>();
+            FileTypes = new HashSet<FileType>();
         }
 
         public int Id { get; set; }
@@ -15,6 +15,6 @@ namespace App.Infrastructures.Database.SqlServer.Entities
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
 
-        public virtual ICollection<Comment> Comments { get; set; }
+        public virtual ICollection<FileType> FileTypes { get; set; }
     }
 }
