+++ App.Domain.Core/BaseData-Aggregate/Entities/Model.cs
@@ -1,17 +1,19 @@
-﻿using System;
+﻿using App.Domain.Core.Product_Aggregate.Entities;
+using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.BaseData_Aggregate.Entities
 {
-    public partial class Brand
+    public class Model
     {
-        public Brand()
+        public Model()
         {
             Products = new HashSet<Product>();
         }
 
         public int Id { get; set; }
-        public int DisplayOrder { get; set; }
+        public int ParentModelId { get; set; }
+        public int BrandId { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
