+++ App.Domain.Core/BaseData-Aggregate/Entities/User.cs
@@ -1,17 +1,16 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.BaseData_Aggregate.Entities
 {
-    public partial class Status
+    public class User
     {
-        public Status()
+        public User()
         {
             Comments = new HashSet<Comment>();
         }
 
         public int Id { get; set; }
-        public bool ForComment { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
