+++ App.Domain.Core/Product-Aggregate/Entities/Brand.cs
@@ -1,19 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product_Aggregate.Entities
 {
-    public partial class Category
+    public class Brand
     {
-        public Category()
+        public Brand()
         {
             Products = new HashSet<Product>();
         }
 
         public int Id { get; set; }
-        public bool IsActive { get; set; }
         public int DisplayOrder { get; set; }
-        public int ParentCagetoryId { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
