+++ App.Domain.Core/Product-Aggregate/Entities/Category.cs
@@ -1,18 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product_Aggregate.Entities
 {
-    public partial class Model
+    public class Category
     {
-        public Model()
+        public Category()
         {
             Products = new HashSet<Product>();
         }
 
         public int Id { get; set; }
-        public int ParentModelId { get; set; }
-        public int BrandId { get; set; }
+        public bool IsActive { get; set; }
+        public int DisplayOrder { get; set; }
+        public int ParentCagetoryId { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
