+++ App.Domain.Core/Product-Aggregate/Entities/ProductColor.cs
@@ -1,19 +1,19 @@
-﻿using System;
+﻿using App.Domain.Core.BaseData_Aggregate.Entities;
+using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product_Aggregate.Entities
 {
-    public partial class ProductTag
+    public class ProductColor
     {
         public int Id { get; set; }
         public int ProductId { get; set; }
-        public int TagId { get; set; }
-        public string Value { get; set; } = null!;
+        public int ColorId { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
 
+        public virtual Color Color { get; set; } = null!;
         public virtual Product Product { get; set; } = null!;
-        public virtual Tag Tag { get; set; } = null!;
     }
 }
