+++ App.Domain.Core/Product-Aggregate/Entities/ProductFile.cs
@@ -1,17 +1,19 @@
-﻿using System;
+﻿using App.Domain.Core.BaseData_Aggregate.Entities;
+using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product_Aggregate.Entities
 {
-    public partial class ProductView
+    public class ProductFile
     {
         public int Id { get; set; }
         public int ProductId { get; set; }
-        public DateTime ViewTime { get; set; }
+        public int FileTypeId { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
 
+        public virtual FileType FileType { get; set; } = null!;
         public virtual Product Product { get; set; } = null!;
     }
 }
