+++ App.Domain.Core/Product-Aggregate/Entities/ProductView.cs
@@ -1,18 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product_Aggregate.Entities
 {
-    public partial class ProductFile
+    public class ProductView
     {
         public int Id { get; set; }
         public int ProductId { get; set; }
-        public int FileTypeId { get; set; }
+        public DateTime ViewTime { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
 
-        public virtual FileType FileType { get; set; } = null!;
         public virtual Product Product { get; set; } = null!;
     }
 }
