+++ App.Domain.Core/Product-Aggregate/Entities/Tag.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product_Aggregate.Entities
 {
-    public partial class Tag
+    public class Tag
     {
         public Tag()
         {
