+++ App.Domain.Services/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace App.Domain.Services
+{
+    public class Class1
+    {
+
+    }
+}
