+++ dotnet-eshop.sln
@@ -5,15 +5,27 @@ VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{A340F267-3318-4AE1-BF7B-457C88F05385}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "01-Infrastructures", "01-Infrastructures", "{BD520F99-027D-4B73-ABB5-77BCD6C30AA5}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "02-Infrastructures", "02-Infrastructures", "{BD520F99-027D-4B73-ABB5-77BCD6C30AA5}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "02-EndPoints", "02-EndPoints", "{54355E0A-CEFB-4327-869F-BA576EE85CCC}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "03-EndPoints", "03-EndPoints", "{54355E0A-CEFB-4327-869F-BA576EE85CCC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Infrastructures.Database.SqlServer", "src\App.Infrastructures.Database.SqlServer\App.Infrastructures.Database.SqlServer.csproj", "{AC0076CE-0A1D-4402-BB1D-4BA8ABC7BCF6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "App.Infrastructures.Database.SqlServer", "src\App.Infrastructures.Database.SqlServer\App.Infrastructures.Database.SqlServer.csproj", "{AC0076CE-0A1D-4402-BB1D-4BA8ABC7BCF6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.EndPoints.Mvc.AdminUI", "src\App.EndPoints.Mvc.AdminUI\App.EndPoints.Mvc.AdminUI.csproj", "{0BE906C9-DE0E-41B9-974D-9DFFB9C72064}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "App.EndPoints.Mvc.AdminUI", "src\App.EndPoints.Mvc.AdminUI\App.EndPoints.Mvc.AdminUI.csproj", "{0BE906C9-DE0E-41B9-974D-9DFFB9C72064}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.EndPoints.Mvc.ShopUI", "src\App.EndPoints.Mvc.ShopUI\App.EndPoints.Mvc.ShopUI.csproj", "{CBF8A57D-6529-414C-9C9B-864A77961283}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "App.EndPoints.Mvc.ShopUI", "src\App.EndPoints.Mvc.ShopUI\App.EndPoints.Mvc.ShopUI.csproj", "{CBF8A57D-6529-414C-9C9B-864A77961283}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "01-Domain", "01-Domain", "{AC202808-9B90-47F8-99C4-737D881DEEB4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{867EF547-41C0-461C-88DC-CB2D4F4AB45D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{6DC771DD-A4D1-49B0-9236-F484EBA24D58}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Domain.Core", "App.Domain.Core\App.Domain.Core.csproj", "{573C3F25-45BD-4084-81B7-6859134B771B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Domain.Services", "App.Domain.Services\App.Domain.Services.csproj", "{C2C3FCA4-9D88-4DBD-BA69-98B073BD2360}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Domain.AppServices", "App.Domain.AppServices\App.Domain.AppServices.csproj", "{0D070F58-5E89-489B-AE23-502887EFEBC0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +45,18 @@ Global
 		{CBF8A57D-6529-414C-9C9B-864A77961283}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CBF8A57D-6529-414C-9C9B-864A77961283}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CBF8A57D-6529-414C-9C9B-864A77961283}.Release|Any CPU.Build.0 = Release|Any CPU
+		{573C3F25-45BD-4084-81B7-6859134B771B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{573C3F25-45BD-4084-81B7-6859134B771B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{573C3F25-45BD-4084-81B7-6859134B771B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{573C3F25-45BD-4084-81B7-6859134B771B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C2C3FCA4-9D88-4DBD-BA69-98B073BD2360}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C2C3FCA4-9D88-4DBD-BA69-98B073BD2360}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C2C3FCA4-9D88-4DBD-BA69-98B073BD2360}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C2C3FCA4-9D88-4DBD-BA69-98B073BD2360}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0D070F58-5E89-489B-AE23-502887EFEBC0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0D070F58-5E89-489B-AE23-502887EFEBC0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0D070F58-5E89-489B-AE23-502887EFEBC0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0D070F58-5E89-489B-AE23-502887EFEBC0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -43,6 +67,12 @@ Global
 		{AC0076CE-0A1D-4402-BB1D-4BA8ABC7BCF6} = {BD520F99-027D-4B73-ABB5-77BCD6C30AA5}
 		{0BE906C9-DE0E-41B9-974D-9DFFB9C72064} = {54355E0A-CEFB-4327-869F-BA576EE85CCC}
 		{CBF8A57D-6529-414C-9C9B-864A77961283} = {54355E0A-CEFB-4327-869F-BA576EE85CCC}
+		{AC202808-9B90-47F8-99C4-737D881DEEB4} = {A340F267-3318-4AE1-BF7B-457C88F05385}
+		{867EF547-41C0-461C-88DC-CB2D4F4AB45D} = {AC202808-9B90-47F8-99C4-737D881DEEB4}
+		{6DC771DD-A4D1-49B0-9236-F484EBA24D58} = {AC202808-9B90-47F8-99C4-737D881DEEB4}
+		{573C3F25-45BD-4084-81B7-6859134B771B} = {867EF547-41C0-461C-88DC-CB2D4F4AB45D}
+		{C2C3FCA4-9D88-4DBD-BA69-98B073BD2360} = {6DC771DD-A4D1-49B0-9236-F484EBA24D58}
+		{0D070F58-5E89-489B-AE23-502887EFEBC0} = {6DC771DD-A4D1-49B0-9236-F484EBA24D58}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E33D7C88-1A57-4CA3-AE1A-7AF235D0924A}
