+++ src/App.EndPoints.Mvc.AdminUI/Controllers/ColorsController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using App.Infrastructures.Database.SqlServer.Repositories.Contracts;
 using App.EndPoints.Mvc.AdminUI.ViewModels;
+using App.Domain.Core.BaseData_Aggregate.Entities;
 
 namespace App.EndPoints.Mvc.AdminUI.Controllers
 {
@@ -32,7 +33,7 @@ namespace App.EndPoints.Mvc.AdminUI.Controllers
         
         public IActionResult SubmitCreatedColor(ColorViewModel model)
         {
-            _colorRepository.Create(new Infrastructures.Database.SqlServer.Entities.Color
+            _colorRepository.Create(new Color
             {
                 Code = model.Code,
                 Name = model.Name
@@ -52,7 +53,7 @@ namespace App.EndPoints.Mvc.AdminUI.Controllers
         }
 
         [HttpPost]
-        public IActionResult Create(App.Infrastructures.Database.SqlServer.Entities.Color model)
+        public IActionResult Create(Color model)
         {
             _colorRepository.Create(model);
             return RedirectToAction("Index");
@@ -66,7 +67,7 @@ namespace App.EndPoints.Mvc.AdminUI.Controllers
             return View();
         }
         [HttpPost]
-        public IActionResult Update(App.Infrastructures.Database.SqlServer.Entities.Color model)
+        public IActionResult Update(Color model)
         {
             _colorRepository.Edit(model);
             return RedirectToAction("Index");
