+++ src/App.EndPoints.Mvc.AdminUI/Views/Brand/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model List<App.Infrastructures.Database.SqlServer.Entities.Brand>
+﻿@model List<App.Domain.Core.Product_Aggregate.Entities.Brand>
 @{
 	ViewData["Title"] = "لیست برند‌ها";
 	Model.OrderBy(x=>x.DisplayOrder);
