+++ src/App.EndPoints.Mvc.AdminUI/Views/Colors/Index.cshtml
@@ -1,6 +1,6 @@
 ﻿
 
-@model List<App.Infrastructures.Database.SqlServer.Entities.Color>
+@model List<App.Domain.Core.BaseData_Aggregate.Entities.Color>
 
 @{
     ViewData["Title"] = "Home Page";
