+++ src/App.EndPoints.Mvc.AdminUI/Views/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@using App.Infrastructures.Database.SqlServer.Entities;
+﻿@using App.Domain.Core.Product_Aggregate.Entities;
 @model List<Product>
 <div class="container p-3">
     <div class="row pt-4">
