+++ src/App.EndPoints.Mvc.AdminUI/Views/Product/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@using App.Infrastructures.Database.SqlServer.Entities;
+﻿@using App.Domain.Core.Product_Aggregate.Entities;
 @model Product
 
 <form method="post" asp-controller="Product" asp-action="DeletePost">
