+++ src/App.EndPoints.Mvc.AdminUI/Views/Product/Edit.cshtml
@@ -1,5 +1,5 @@
 ﻿
-@using App.Infrastructures.Database.SqlServer.Entities;
+@using App.Domain.Core.Product_Aggregate.Entities;
 @model Product
 
 <form method="post"  asp-controller="Product" asp-action="Edit">
