+++ src/App.Infrastructures.Database.SqlServer/Repositories/CategoryRepository.cs
@@ -1,5 +1,5 @@
 ﻿using App.Infrastructures.Database.SqlServer.Data;
-using App.Infrastructures.Database.SqlServer.Entities;
+using App.Domain.Core.Product_Aggregate.Entities;
 using App.Infrastructures.Database.SqlServer.Repositories.Contracts;
 
 namespace App.Infrastructures.Database.SqlServer.Repositories
