+++ src/App.Infrastructures.Database.SqlServer/Repositories/ColorEfRepository.cs
@@ -1,5 +1,5 @@
 ﻿using App.Infrastructures.Database.SqlServer.Data;
-using App.Infrastructures.Database.SqlServer.Entities;
+using App.Domain.Core.Product_Aggregate.Entities;
 using App.Infrastructures.Database.SqlServer.Repositories.Contracts;
 using Microsoft.EntityFrameworkCore;
 using System;
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using App.Domain.Core.BaseData_Aggregate.Entities;
 
 namespace App.Infrastructures.Database.SqlServer.Repositories
 {
@@ -49,5 +50,7 @@ namespace App.Infrastructures.Database.SqlServer.Repositories
         {
             return _eshop.Colors.First(p => p.Id == id);
         }
+
+       
     }
 }
