+++ src/App.Infrastructures.Database.SqlServer/Repositories/ColorInMemoryRepository.cs
@@ -1,4 +1,5 @@
-﻿using App.Infrastructures.Database.SqlServer.Data;
+﻿using App.Domain.Core.BaseData_Aggregate.Entities;
+using App.Infrastructures.Database.SqlServer.Data;
 using App.Infrastructures.Database.SqlServer.Repositories.Contracts;
 using Microsoft.EntityFrameworkCore;
 
@@ -14,12 +15,12 @@ namespace App.Infrastructures.Database.SqlServer.Repositories
             _eshop = appDbContext;
         }
 
-        public void Create(Entities.Color color)
+        public void Create(Color color)
         {
             _eshop.Colors.Add(color);
             _eshop.SaveChanges();
         }
-        public void Edit(Entities.Color model)
+        public void Edit(Color model)
         {
             var color = _eshop.Colors.First(p => p.Id == model.Id);
             color.Name = model.Name;
@@ -35,13 +36,17 @@ namespace App.Infrastructures.Database.SqlServer.Repositories
         }
 
 
-        public List<Entities.Color> GetAll()
+        public List<Color> GetAll()
         {
             return _eshop.Colors.Include(b => b.ProductColors).ToList();
         }
-        public Entities.Color GetById(int id)
+        public Color GetById(int id)
         {
             return _eshop.Colors.First(p => p.Id == id);
         }
+
+       
+        
+        
     }
 }
