+++ src/App.Infrastructures.Database.SqlServer/Repositories/Contracts/ICategoryRepository.cs
@@ -1,4 +1,4 @@
-﻿using App.Infrastructures.Database.SqlServer.Entities;
+﻿using App.Domain.Core.Product_Aggregate.Entities;
 
 namespace App.Infrastructures.Database.SqlServer.Repositories.Contracts
 {
