+++ src/App.Infrastructures.Database.SqlServer/Repositories/Contracts/IModelRepository.cs
@@ -1,4 +1,4 @@
-﻿using App.Infrastructures.Database.SqlServer.Entities;
+﻿using App.Domain.Core.BaseData_Aggregate.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
