--- src/App.Infrastructures.Database.SqlServer/Repositories/IColorRepository.cs
@@ -1,13 +1,0 @@
-﻿using App.Infrastructures.Database.SqlServer.Entities;
-
-namespace App.Infrastructures.Database.SqlServer.Ripository
-{
-    public interface IColorRepository
-    {
-        void Create(Color color);
-        void Delete(int id);
-        void Edit(Color model);
-        List<Color> GetAll();
-        Color GetById(int id);
-    }
-}
